test(bayn): run authority migration proof in postgres suite
